back/requirements.txt
alembic==1.5.1
aniso8601==9.0.1
asgiref==3.4.1
attrs==21.4.0
bcrypt==3.2.0
captcha==0.4
certifi==2021.5.30
cffi==1.15.0
chardet==4.0.0
click==8.0.4
dataclasses==0.8
Django==3.2.12
django-recaptcha==2.0.6
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-JWT-Extended==3.25.0
Flask-Migrate==2.6.0
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-Sockets==0.2.1
Flask-SQLAlchemy==2.4.4
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.2
idna==2.10
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.0.0
Mako==1.1.6
MarkupSafe==2.0.1
Pillow==8.1.0
pycparser==2.21
pycryptodome==3.9.9
PyJWT==1.7.1
PyMySQL==1.0.2
pyrsistent==0.18.0
python-dateutil==2.8.2
python-editor==1.0.4
pytz==2022.1
requests==2.25.1
six==1.16.0
SQLAlchemy==1.3.22
sqlparse==0.4.2
typing_extensions==4.1.1
urllib3==1.26.9
Werkzeug==0.16.0
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.4.0
